feat: implement lazy-loaded globe canvas in authentication forms
- Introduced a new GlobeCanvas component for interactive globe visualization, enhancing user engagement across the login, registration, and forgot password forms.
- Replaced previous globe rendering logic with a lazy-loaded component to improve performance and maintainability.
- Updated form layouts to utilize the new GlobeCanvas, ensuring consistent styling and responsiveness across all authentication components.
